backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
botocore==1.40.17
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
executing==2.2.0
fastapi==0.119.0
fastjsonschema==2.21.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fqdn==1.5.1
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
huepy==1.2.1
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
livereload==2.7.1
Markdown==3.8.2
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.4
opencv-python==4.12.0.88
opt_einsum==3.4.0
optional-django==0.3.0
optree==0.17.0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
react==4.3.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.24.0
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.3
setuptools==78.1.0
simple-parsing==0.1.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
zipp==3.23.0
